+++ app/com/lynxanalytics/biggraph/graph_operations/NeuralNetwork.scala
@@ -61,6 +61,7 @@ case class PredictViaNNOnGraphV1(
     gradientCheckOn: Boolean,
     networkLayout: String) extends TypedMetaGraphOp[Input, Output] {
   @transient override lazy val inputs = new Input(featureCount)
+  override val isHeavy = true
   def outputMeta(instance: MetaGraphOperationInstance) = new Output()(instance, inputs)
   override def toJson = Json.obj(
     "featureCount" -> featureCount,
@@ -146,7 +147,7 @@ case class PredictViaNNOnGraphV1(
     seed: Int): (Seq[ID], Map[ID, Seq[ID]], Seq[(ID, (Option[Double], Array[Double]))]) = {
     val data = dataIterator.toSeq
     val vertices = data.map(_._1)
-    if (selectionRadius < 0) { // Return the whole graph, for testing.
+    if (selectionRadius <= 0) { // Return the whole graph, for testing.
       (vertices, edgeLists, data)
     } else {
       val random = new util.Random(seed)
@@ -341,7 +342,7 @@ case class PredictViaNNOnGraphV1(
             val rows = values.rows
             val cols = values.cols
             for (row <- 0 until rows; col <- 0 until cols) yield (s"$name $row $col", values(row, col))
-        }.toMap
+        }
       }
 
       var gradientsOK = true
@@ -358,10 +359,9 @@ case class PredictViaNNOnGraphV1(
               println(s"Gradient check fails on $name, backprop grad = $otherValue, approximated grad = $value")
               gradientsOK = false
             }
-
-            (name, relativeError)
-        }.toMap
+        }
       }
+      log.debug(s"relativeErrors: $relativeErrors")
       gradientsOK
     }