+++ tools/e2e_test.sh
@@ -2,5 +2,4 @@
 # Run this with with_lk.sh.
 
 cd $(dirname $0)/..
-tools/test_interface.sh
 gulp --cwd web test